less bad way of doing it
